For 30013, I have changed the description to “Patient advises that the service is no longer needed.”
I have corrected the Definition for “fatal-error”, to match HL7 valueSet response-code definition, “The message was rejected because of a problem with the content. There is no point in re-sending without change. The response narrative SHALL describe the issue.”

Similarly for the Definition of “informational” I have corrected this to match the HL7 issue-type ValueSet definition, “ A message unrelated to the processing success of the completed operation (examples of the latter include things like reminders of password expiry, system maintenance times, etc.).”